pyproject.toml
[tool.poetry]
name = "nas-test"
version = "0.1.0"
description = ""
authors = ["Joe Burris"]
readme = "README.md"
packages = [{include = "nas_test"}]

[tool.poetry.dependencies]
python = "^3.8"
nni = "^2.10.1"
torch = "1.10.0"
pytorch-lightning = "1.6.1"
torchvision = "0.11.1"
matplotlib = "^3.7.2"
numba = "^0.57.1"
pydantic = ">=1.7.4,<2.2.0"
fastapi = ">=0.92.0,<0.100.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
